pyproject.toml
[tool.poetry]
name = "extraction"
version = "0.0.1"
readme = "README.md"
authors = ["Bek"]
description = "Extraction"
packages = [{include = "extraction"}]

[tool.poetry.dependencies]
python = "^3.12"
langchain = "^0.3.19"
langchain-community = "^0.3.18"
openai = "^1.65.2"
pypdf = "^5.0.1"
openpyxl = "^3.1.5"
fastapi = "^0.110.1"
uvicorn = "^0.21.0"
python-dotenv = "^0.21.1"
pydantic = "2.7.4"


[tool.poetry.scripts]
start = "extraction.server:start"

[tool.poetry.group.test.dependencies]
pytest = "7.2.1"
mockito = "^1.4.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
